[ALOY-1074] Enable developers to specify Backbone version in config.json for 1_6_X
