+++ Alloy/lib/alloy/backbone/0.9.2/backbone.js
@@ -1,0 +1,1431 @@
+//     Backbone.js 0.9.2
+
+//     (c) 2010-2012 Jeremy Ashkenas, DocumentCloud Inc.
+//     Backbone may be freely distributed under the MIT license.
+//     For all details and documentation:
+//     http://backbonejs.org
+
+(function(){
+
+// Initial Setup
+// -------------
+
+// Save a reference to the global object (`window` in the browser, `global`
+// on the server).
+var root = this;
+
+// Save the previous value of the `Backbone` variable, so that it can be
+// restored later on, if `noConflict` is used.
+var previousBackbone = root.Backbone;
+
+// Create a local reference to slice/splice.
+var slice = Array.prototype.slice;
+var splice = Array.prototype.splice;
+
+// The top-level namespace. All public Backbone classes and modules will
+// be attached to this. Exported for both CommonJS and the browser.
+var Backbone;
+if (typeof exports !== 'undefined') {
+	Backbone = exports;
+} else {
+	Backbone = root.Backbone = {};
+}
+
+// Current version of the library. Keep in sync with `package.json`.
+Backbone.VERSION = '0.9.2';
+
+// Require Underscore, if we're on the server, and it's not already present.
+var _ = root._;
+if (!_ && (typeof require !== 'undefined')) _ = require('alloy/underscore');
+
+// For Backbone's purposes, jQuery, Zepto, or Ender owns the `$` variable.
+var $ = root.jQuery || root.Zepto || root.ender;
+
+// Set the JavaScript library that will be used for DOM manipulation and
+// Ajax calls (a.k.a. the `$` variable). By default Backbone will use: jQuery,
+// Zepto, or Ender; but the `setDomLibrary()` method lets you inject an
+// alternate JavaScript library (or a mock library for testing your views
+// outside of a browser).
+Backbone.setDomLibrary = function(lib) {
+	$ = lib;
+};
+
+// Runs Backbone.js in *noConflict* mode, returning the `Backbone` variable
+// to its previous owner. Returns a reference to this Backbone object.
+Backbone.noConflict = function() {
+	root.Backbone = previousBackbone;
+	return this;
+};
+
+// Turn on `emulateHTTP` to support legacy HTTP servers. Setting this option
+// will fake `"PUT"` and `"DELETE"` requests via the `_method` parameter and
+// set a `X-Http-Method-Override` header.
+Backbone.emulateHTTP = false;
+
+// Turn on `emulateJSON` to support legacy servers that can't deal with direct
+// `application/json` requests ... will encode the body as
+// `application/x-www-form-urlencoded` instead and will send the model in a
+// form param named `model`.
+Backbone.emulateJSON = false;
+
+// Backbone.Events
+// -----------------
+
+// Regular expression used to split event strings
+var eventSplitter = /\s+/;
+
+// A module that can be mixed in to *any object* in order to provide it with
+// custom events. You may bind with `on` or remove with `off` callback functions
+// to an event; trigger`-ing an event fires all callbacks in succession.
+//
+//     var object = {};
+//     _.extend(object, Backbone.Events);
+//     object.on('expand', function(){ alert('expanded'); });
+//     object.trigger('expand');
+//
+var Events = Backbone.Events = {
+
+	// Bind one or more space separated events, `events`, to a `callback`
+	// function. Passing `"all"` will bind the callback to all events fired.
+	on: function(events, callback, context) {
+
+	var calls, event, node, tail, list;
+	if (!callback) return this;
+	events = events.split(eventSplitter);
+	calls = this._callbacks || (this._callbacks = {});
+
+	// Create an immutable callback list, allowing traversal during
+	// modification.  The tail is an empty object that will always be used
+	// as the next node.
+	while (event = events.shift()) {
+		list = calls[event];
+		node = list ? list.tail : {};
+		node.next = tail = {};
+		node.context = context;
+		node.callback = callback;
+		calls[event] = {tail: tail, next: list ? list.next : node};
+	}
+
+	return this;
+	},
+
+	// Remove one or many callbacks. If `context` is null, removes all callbacks
+	// with that function. If `callback` is null, removes all callbacks for the
+	// event. If `events` is null, removes all bound callbacks for all events.
+	off: function(events, callback, context) {
+	var event, calls, node, tail, cb, ctx;
+
+	// No events, or removing *all* events.
+	if (!(calls = this._callbacks)) return;
+	if (!(events || callback || context)) {
+		delete this._callbacks;
+		return this;
+	}
+
+	// Loop through the listed events and contexts, splicing them out of the
+	// linked list of callbacks if appropriate.
+	events = events ? events.split(eventSplitter) : _.keys(calls);
+	while (event = events.shift()) {
+		node = calls[event];
+		delete calls[event];
+		if (!node || !(callback || context)) continue;
+		// Create a new list, omitting the indicated callbacks.
+		tail = node.tail;
+		while ((node = node.next) !== tail) {
+		cb = node.callback;
+		ctx = node.context;
+		if ((callback && cb !== callback) || (context && ctx !== context)) {
+			this.on(event, cb, ctx);
+		}
+		}
+	}
+
+	return this;
+	},
+
+	// Trigger one or many events, firing all bound callbacks. Callbacks are
+	// passed the same arguments as `trigger` is, apart from the event name
+	// (unless you're listening on `"all"`, which will cause your callback to
+	// receive the true name of the event as the first argument).
+	trigger: function(events) {
+	var event, node, calls, tail, args, all, rest;
+	if (!(calls = this._callbacks)) return this;
+	all = calls.all;
+	events = events.split(eventSplitter);
+	rest = slice.call(arguments, 1);
+
+	// For each event, walk through the linked list of callbacks twice,
+	// first to trigger the event, then to trigger any `"all"` callbacks.
+	while (event = events.shift()) {
+		if (node = calls[event]) {
+		tail = node.tail;
+		while ((node = node.next) !== tail) {
+			node.callback.apply(node.context || this, rest);
+		}
+		}
+		if (node = all) {
+		tail = node.tail;
+		args = [event].concat(rest);
+		while ((node = node.next) !== tail) {
+			node.callback.apply(node.context || this, args);
+		}
+		}
+	}
+
+	return this;
+	}
+
+};
+
+// Aliases for backwards compatibility.
+Events.bind   = Events.on;
+Events.unbind = Events.off;
+
+// Backbone.Model
+// --------------
+
+// Create a new model, with defined attributes. A client id (`cid`)
+// is automatically generated and assigned for you.
+var Model = Backbone.Model = function(attributes, options) {
+	var defaults;
+	attributes || (attributes = {});
+	if (options && options.parse) attributes = this.parse(attributes);
+	if (defaults = getValue(this, 'defaults')) {
+	attributes = _.extend({}, defaults, attributes);
+	}
+	if (options && options.collection) this.collection = options.collection;
+	this.attributes = {};
+	this._escapedAttributes = {};
+	this.cid = _.uniqueId('c');
+	this.changed = {};
+	this._silent = {};
+	this._pending = {};
+	this.set(attributes, {silent: true});
+	// Reset change tracking.
+	this.changed = {};
+	this._silent = {};
+	this._pending = {};
+	this._previousAttributes = _.clone(this.attributes);
+	this.initialize.apply(this, arguments);
+};
+
+// Attach all inheritable methods to the Model prototype.
+_.extend(Model.prototype, Events, {
+
+	// A hash of attributes whose current and previous value differ.
+	changed: null,
+
+	// A hash of attributes that have silently changed since the last time
+	// `change` was called.  Will become pending attributes on the next call.
+	_silent: null,
+
+	// A hash of attributes that have changed since the last `'change'` event
+	// began.
+	_pending: null,
+
+	// The default name for the JSON `id` attribute is `"id"`. MongoDB and
+	// CouchDB users may want to set this to `"_id"`.
+	idAttribute: 'id',
+
+	// Initialize is an empty function by default. Override it with your own
+	// initialization logic.
+	initialize: function(){},
+
+	// Return a copy of the model's `attributes` object.
+	toJSON: function(options) {
+	return _.clone(this.attributes);
+	},
+
+	// Get the value of an attribute.
+	get: function(attr) {
+	return this.attributes[attr];
+	},
+
+	// Get the HTML-escaped value of an attribute.
+	escape: function(attr) {
+	var html;
+	if (html = this._escapedAttributes[attr]) return html;
+	var val = this.get(attr);
+	return this._escapedAttributes[attr] = _.escape(val == null ? '' : '' + val);
+	},
+
+	// Returns `true` if the attribute contains a value that is not null
+	// or undefined.
+	has: function(attr) {
+	return this.get(attr) != null;
+	},
+
+	// Set a hash of model attributes on the object, firing `"change"` unless
+	// you choose to silence it.
+	set: function(key, value, options) {
+	var attrs, attr, val;
+
+	// Handle both
+	if (_.isObject(key) || key == null) {
+		attrs = key;
+		options = value;
+	} else {
+		attrs = {};
+		attrs[key] = value;
+	}
+
+	// Extract attributes and options.
+	options || (options = {});
+	if (!attrs) return this;
+	if (attrs instanceof Model) attrs = attrs.attributes;
+	if (options.unset) for (attr in attrs) attrs[attr] = void 0;
+
+	// Run validation.
+	if (!this._validate(attrs, options)) return false;
+
+	// Check for changes of `id`.
+	if (this.idAttribute in attrs) this.id = attrs[this.idAttribute];
+
+	var changes = options.changes = {};
+	var now = this.attributes;
+	var escaped = this._escapedAttributes;
+	var prev = this._previousAttributes || {};
+
+	// For each `set` attribute...
+	for (attr in attrs) {
+		val = attrs[attr];
+
+		// If the new and current value differ, record the change.
+		if (!_.isEqual(now[attr], val) || (options.unset && _.has(now, attr))) {
+		delete escaped[attr];
+		(options.silent ? this._silent : changes)[attr] = true;
+		}
+
+		// Update or delete the current value.
+		options.unset ? delete now[attr] : now[attr] = val;
+
+		// If the new and previous value differ, record the change.  If not,
+		// then remove changes for this attribute.
+		if (!_.isEqual(prev[attr], val) || (_.has(now, attr) != _.has(prev, attr))) {
+		this.changed[attr] = val;
+		if (!options.silent) this._pending[attr] = true;
+		} else {
+		delete this.changed[attr];
+		delete this._pending[attr];
+		}
+	}
+
+	// Fire the `"change"` events.
+	if (!options.silent) this.change(options);
+	return this;
+	},
+
+	// Remove an attribute from the model, firing `"change"` unless you choose
+	// to silence it. `unset` is a noop if the attribute doesn't exist.
+	unset: function(attr, options) {
+	(options || (options = {})).unset = true;
+	return this.set(attr, null, options);
+	},
+
+	// Clear all attributes on the model, firing `"change"` unless you choose
+	// to silence it.
+	clear: function(options) {
+	(options || (options = {})).unset = true;
+	return this.set(_.clone(this.attributes), options);
+	},
+
+	// Fetch the model from the server. If the server's representation of the
+	// model differs from its current attributes, they will be overriden,
+	// triggering a `"change"` event.
+	fetch: function(options) {
+	options = options ? _.clone(options) : {};
+	var model = this;
+	var success = options.success;
+	options.success = function(resp, status, xhr) {
+		if (!model.set(model.parse(resp, xhr), options)) return false;
+		if (success) success(model, resp);
+	};
+	options.error = Backbone.wrapError(options.error, model, options);
+	return (this.sync || Backbone.sync).call(this, 'read', this, options);
+	},
+
+	// Set a hash of model attributes, and sync the model to the server.
+	// If the server returns an attributes hash that differs, the model's
+	// state will be `set` again.
+	save: function(key, value, options) {
+	var attrs, current;
+
+	// Handle both `("key", value)` and `({key: value})` -style calls.
+	if (_.isObject(key) || key == null) {
+		attrs = key;
+		options = value;
+	} else {
+		attrs = {};
+		attrs[key] = value;
+	}
+	options = options ? _.clone(options) : {};
+
+	// If we're "wait"-ing to set changed attributes, validate early.
+	if (options.wait) {
+		if (!this._validate(attrs, options)) return false;
+		current = _.clone(this.attributes);
+	}
+
+	// Regular saves `set` attributes before persisting to the server.
+	var silentOptions = _.extend({}, options, {silent: true});
+	if (attrs && !this.set(attrs, options.wait ? silentOptions : options)) {
+		return false;
+	}
+
+	// After a successful server-side save, the client is (optionally)
+	// updated with the server-side state.
+	var model = this;
+	var success = options.success;
+	options.success = function(resp, status, xhr) {
+		var serverAttrs = model.parse(resp, xhr);
+		if (options.wait) {
+		delete options.wait;
+		serverAttrs = _.extend(attrs || {}, serverAttrs);
+		}
+		if (!model.set(serverAttrs, options)) return false;
+		if (success) {
+		success(model, resp);
+		} else {
+		model.trigger('sync', model, resp, options);
+		}
+	};
+
+	// Finish configuring and sending the Ajax request.
+	options.error = Backbone.wrapError(options.error, model, options);
+	var method = this.isNew() ? 'create' : 'update';
+	var xhr = (this.sync || Backbone.sync).call(this, method, this, options);
+	if (options.wait) this.set(current, silentOptions);
+	return xhr;
+	},
+
+	// Destroy this model on the server if it was already persisted.
+	// Optimistically removes the model from its collection, if it has one.
+	// If `wait: true` is passed, waits for the server to respond before removal.
+	destroy: function(options) {
+	options = options ? _.clone(options) : {};
+	var model = this;
+	var success = options.success;
+
+	var triggerDestroy = function() {
+		model.trigger('destroy', model, model.collection, options);
+	};
+
+	if (this.isNew()) {
+		triggerDestroy();
+		return false;
+	}
+
+	options.success = function(resp) {
+		if (options.wait) triggerDestroy();
+		if (success) {
+		success(model, resp);
+		} else {
+		model.trigger('sync', model, resp, options);
+		}
+	};
+
+	options.error = Backbone.wrapError(options.error, model, options);
+	var xhr = (this.sync || Backbone.sync).call(this, 'delete', this, options);
+	if (!options.wait) triggerDestroy();
+	return xhr;
+	},
+
+	// Default URL for the model's representation on the server -- if you're
+	// using Backbone's restful methods, override this to change the endpoint
+	// that will be called.
+	url: function() {
+	var base = getValue(this, 'urlRoot') || getValue(this.collection, 'url') || urlError();
+	if (this.isNew()) return base;
+	return base + (base.charAt(base.length - 1) == '/' ? '' : '/') + encodeURIComponent(this.id);
+	},
+
+	// **parse** converts a response into the hash of attributes to be `set` on
+	// the model. The default implementation is just to pass the response along.
+	parse: function(resp, xhr) {
+	return resp;
+	},
+
+	// Create a new model with identical attributes to this one.
+	clone: function() {
+	return new this.constructor(this.attributes);
+	},
+
+	// A model is new if it has never been saved to the server, and lacks an id.
+	isNew: function() {
+	return this.id == null;
+	},
+
+	// Call this method to manually fire a `"change"` event for this model and
+	// a `"change:attribute"` event for each changed attribute.
+	// Calling this will cause all objects observing the model to update.
+	change: function(options) {
+	options || (options = {});
+	var changing = this._changing;
+	this._changing = true;
+
+	// Silent changes become pending changes.
+	for (var attr in this._silent) this._pending[attr] = true;
+
+	// Silent changes are triggered.
+	var changes = _.extend({}, options.changes, this._silent);
+	this._silent = {};
+	for (var attr in changes) {
+		this.trigger('change:' + attr, this, this.get(attr), options);
+	}
+	if (changing) return this;
+
+	// Continue firing `"change"` events while there are pending changes.
+	while (!_.isEmpty(this._pending)) {
+		this._pending = {};
+		this.trigger('change', this, options);
+		// Pending and silent changes still remain.
+		for (var attr in this.changed) {
+		if (this._pending[attr] || this._silent[attr]) continue;
+		delete this.changed[attr];
+		}
+		this._previousAttributes = _.clone(this.attributes);
+	}
+
+	this._changing = false;
+	return this;
+	},
+
+	// Determine if the model has changed since the last `"change"` event.
+	// If you specify an attribute name, determine if that attribute has changed.
+	hasChanged: function(attr) {
+	if (!arguments.length) return !_.isEmpty(this.changed);
+	return _.has(this.changed, attr);
+	},
+
+	// Return an object containing all the attributes that have changed, or
+	// false if there are no changed attributes. Useful for determining what
+	// parts of a view need to be updated and/or what attributes need to be
+	// persisted to the server. Unset attributes will be set to undefined.
+	// You can also pass an attributes object to diff against the model,
+	// determining if there *would be* a change.
+	changedAttributes: function(diff) {
+	if (!diff) return this.hasChanged() ? _.clone(this.changed) : false;
+	var val, changed = false, old = this._previousAttributes;
+	for (var attr in diff) {
+		if (_.isEqual(old[attr], (val = diff[attr]))) continue;
+		(changed || (changed = {}))[attr] = val;
+	}
+	return changed;
+	},
+
+	// Get the previous value of an attribute, recorded at the time the last
+	// `"change"` event was fired.
+	previous: function(attr) {
+	if (!arguments.length || !this._previousAttributes) return null;
+	return this._previousAttributes[attr];
+	},
+
+	// Get all of the attributes of the model at the time of the previous
+	// `"change"` event.
+	previousAttributes: function() {
+	return _.clone(this._previousAttributes);
+	},
+
+	// Check if the model is currently in a valid state. It's only possible to
+	// get into an *invalid* state if you're using silent changes.
+	isValid: function() {
+	return !this.validate(this.attributes);
+	},
+
+	// Run validation against the next complete set of model attributes,
+	// returning `true` if all is well. If a specific `error` callback has
+	// been passed, call that instead of firing the general `"error"` event.
+	_validate: function(attrs, options) {
+	if (options.silent || !this.validate) return true;
+	attrs = _.extend({}, this.attributes, attrs);
+	var error = this.validate(attrs, options);
+	if (!error) return true;
+	if (options && options.error) {
+		options.error(this, error, options);
+	} else {
+		this.trigger('error', this, error, options);
+	}
+	return false;
+	}
+
+});
+
+// Backbone.Collection
+// -------------------
+
+// Provides a standard collection class for our sets of models, ordered
+// or unordered. If a `comparator` is specified, the Collection will maintain
+// its models in sort order, as they're added and removed.
+var Collection = Backbone.Collection = function(models, options) {
+	options || (options = {});
+	if (options.model) this.model = options.model;
+	if (options.comparator) this.comparator = options.comparator;
+	this._reset();
+	this.initialize.apply(this, arguments);
+	if (models) this.reset(models, {silent: true, parse: options.parse});
+};
+
+// Define the Collection's inheritable methods.
+_.extend(Collection.prototype, Events, {
+
+	// The default model for a collection is just a **Backbone.Model**.
+	// This should be overridden in most cases.
+	model: Model,
+
+	// Initialize is an empty function by default. Override it with your own
+	// initialization logic.
+	initialize: function(){},
+
+	// The JSON representation of a Collection is an array of the
+	// models' attributes.
+	toJSON: function(options) {
+	return this.map(function(model){ return model.toJSON(options); });
+	},
+
+	// Add a model, or list of models to the set. Pass **silent** to avoid
+	// firing the `add` event for every new model.
+	add: function(models, options) {
+	var i, index, length, model, cid, id, cids = {}, ids = {}, dups = [];
+	options || (options = {});
+	models = _.isArray(models) ? models.slice() : [models];
+
+	// Begin by turning bare objects into model references, and preventing
+	// invalid models or duplicate models from being added.
+	for (i = 0, length = models.length; i < length; i++) {
+		if (!(model = models[i] = this._prepareModel(models[i], options))) {
+		throw new Error("Can't add an invalid model to a collection");
+		}
+		cid = model.cid;
+		id = model.id;
+		if (cids[cid] || this._byCid[cid] || ((id != null) && (ids[id] || this._byId[id]))) {
+		dups.push(i);
+		continue;
+		}
+		cids[cid] = ids[id] = model;
+	}
+
+	// Remove duplicates.
+	i = dups.length;
+	while (i--) {
+		models.splice(dups[i], 1);
+	}
+
+	// Listen to added models' events, and index models for lookup by
+	// `id` and by `cid`.
+	for (i = 0, length = models.length; i < length; i++) {
+		(model = models[i]).on('all', this._onModelEvent, this);
+		this._byCid[model.cid] = model;
+		if (model.id != null) this._byId[model.id] = model;
+	}
+
+	// Insert models into the collection, re-sorting if needed, and triggering
+	// `add` events unless silenced.
+	this.length += length;
+	index = options.at != null ? options.at : this.models.length;
+	splice.apply(this.models, [index, 0].concat(models));
+	if (this.comparator) this.sort({silent: true});
+	if (options.silent) return this;
+	for (i = 0, length = this.models.length; i < length; i++) {
+		if (!cids[(model = this.models[i]).cid]) continue;
+		options.index = i;
+		model.trigger('add', model, this, options);
+	}
+	return this;
+	},
+
+	// Remove a model, or a list of models from the set. Pass silent to avoid
+	// firing the `remove` event for every model removed.
+	remove: function(models, options) {
+	var i, l, index, model;
+	options || (options = {});
+	models = _.isArray(models) ? models.slice() : [models];
+	for (i = 0, l = models.length; i < l; i++) {
+		model = this.getByCid(models[i]) || this.get(models[i]);
+		if (!model) continue;
+		delete this._byId[model.id];
+		delete this._byCid[model.cid];
+		index = this.indexOf(model);
+		this.models.splice(index, 1);
+		this.length--;
+		if (!options.silent) {
+		options.index = index;
+		model.trigger('remove', model, this, options);
+		}
+		this._removeReference(model);
+	}
+	return this;
+	},
+
+	// Add a model to the end of the collection.
+	push: function(model, options) {
+	model = this._prepareModel(model, options);
+	this.add(model, options);
+	return model;
+	},
+
+	// Remove a model from the end of the collection.
+	pop: function(options) {
+	var model = this.at(this.length - 1);
+	this.remove(model, options);
+	return model;
+	},
+
+	// Add a model to the beginning of the collection.
+	unshift: function(model, options) {
+	model = this._prepareModel(model, options);
+	this.add(model, _.extend({at: 0}, options));
+	return model;
+	},
+
+	// Remove a model from the beginning of the collection.
+	shift: function(options) {
+	var model = this.at(0);
+	this.remove(model, options);
+	return model;
+	},
+
+	// Get a model from the set by id.
+	get: function(id) {
+	if (id == null) return void 0;
+	return this._byId[id.id != null ? id.id : id];
+	},
+
+	// Get a model from the set by client id.
+	getByCid: function(cid) {
+	return cid && this._byCid[cid.cid || cid];
+	},
+
+	// Get the model at the given index.
+	at: function(index) {
+	return this.models[index];
+	},
+
+	// Return models with matching attributes. Useful for simple cases of `filter`.
+	where: function(attrs) {
+	if (_.isEmpty(attrs)) return [];
+	return this.filter(function(model) {
+		for (var key in attrs) {
+		if (attrs[key] !== model.get(key)) return false;
+		}
+		return true;
+	});
+	},
+
+	// Force the collection to re-sort itself. You don't need to call this under
+	// normal circumstances, as the set will maintain sort order as each item
+	// is added.
+	sort: function(options) {
+	options || (options = {});
+	if (!this.comparator) throw new Error('Cannot sort a set without a comparator');
+	var boundComparator = _.bind(this.comparator, this);
+	if (this.comparator.length == 1) {
+		this.models = this.sortBy(boundComparator);
+	} else {
+		this.models.sort(boundComparator);
+	}
+	if (!options.silent) this.trigger('reset', this, options);
+	return this;
+	},
+
+	// Pluck an attribute from each model in the collection.
+	pluck: function(attr) {
+	return _.map(this.models, function(model){ return model.get(attr); });
+	},
+
+	// When you have more items than you want to add or remove individually,
+	// you can reset the entire set with a new list of models, without firing
+	// any `add` or `remove` events. Fires `reset` when finished.
+	reset: function(models, options) {
+	models  || (models = []);
+	options || (options = {});
+	for (var i = 0, l = this.models.length; i < l; i++) {
+		this._removeReference(this.models[i]);
+	}
+	this._reset();
+	this.add(models, _.extend({silent: true}, options));
+	if (!options.silent) this.trigger('reset', this, options);
+	return this;
+	},
+
+	// Fetch the default set of models for this collection, resetting the
+	// collection when they arrive. If `add: true` is passed, appends the
+	// models to the collection instead of resetting.
+	fetch: function(options) {
+	options = options ? _.clone(options) : {};
+	if (options.parse === undefined) options.parse = true;
+	var collection = this;
+	var success = options.success;
+	options.success = function(resp, status, xhr) {
+		collection[options.add ? 'add' : 'reset'](collection.parse(resp, xhr), options);
+		if (success) success(collection, resp);
+	};
+	options.error = Backbone.wrapError(options.error, collection, options);
+	return (this.sync || Backbone.sync).call(this, 'read', this, options);
+	},
+
+	// Create a new instance of a model in this collection. Add the model to the
+	// collection immediately, unless `wait: true` is passed, in which case we
+	// wait for the server to agree.
+	create: function(model, options) {
+	var coll = this;
+	options = options ? _.clone(options) : {};
+	model = this._prepareModel(model, options);
+	if (!model) return false;
+	if (!options.wait) coll.add(model, options);
+	var success = options.success;
+	options.success = function(nextModel, resp, xhr) {
+		if (options.wait) coll.add(nextModel, options);
+		if (success) {
+		success(nextModel, resp);
+		} else {
+		nextModel.trigger('sync', model, resp, options);
+		}
+	};
+	model.save(null, options);
+	return model;
+	},
+
+	// **parse** converts a response into a list of models to be added to the
+	// collection. The default implementation is just to pass it through.
+	parse: function(resp, xhr) {
+	return resp;
+	},
+
+	// Proxy to _'s chain. Can't be proxied the same way the rest of the
+	// underscore methods are proxied because it relies on the underscore
+	// constructor.
+	chain: function () {
+	return _(this.models).chain();
+	},
+
+	// Reset all internal state. Called when the collection is reset.
+	_reset: function(options) {
+	this.length = 0;
+	this.models = [];
+	this._byId  = {};
+	this._byCid = {};
+	},
+
+	// Prepare a model or hash of attributes to be added to this collection.
+	_prepareModel: function(model, options) {
+	options || (options = {});
+	if (!(model instanceof Model)) {
+		var attrs = model;
+		options.collection = this;
+		model = new this.model(attrs, options);
+		if (!model._validate(model.attributes, options)) model = false;
+	} else if (!model.collection) {
+		model.collection = this;
+	}
+	return model;
+	},
+
+	// Internal method to remove a model's ties to a collection.
+	_removeReference: function(model) {
+	if (this == model.collection) {
+		delete model.collection;
+	}
+	model.off('all', this._onModelEvent, this);
+	},
+
+	// Internal method called every time a model in the set fires an event.
+	// Sets need to update their indexes when models change ids. All other
+	// events simply proxy through. "add" and "remove" events that originate
+	// in other collections are ignored.
+	_onModelEvent: function(event, model, collection, options) {
+	if ((event == 'add' || event == 'remove') && collection != this) return;
+	if (event == 'destroy') {
+		this.remove(model, options);
+	}
+	if (model && event === 'change:' + model.idAttribute) {
+		delete this._byId[model.previous(model.idAttribute)];
+		this._byId[model.id] = model;
+	}
+	this.trigger.apply(this, arguments);
+	}
+
+});
+
+// Underscore methods that we want to implement on the Collection.
+var methods = ['forEach', 'each', 'map', 'reduce', 'reduceRight', 'find',
+	'detect', 'filter', 'select', 'reject', 'every', 'all', 'some', 'any',
+	'include', 'contains', 'invoke', 'max', 'min', 'sortBy', 'sortedIndex',
+	'toArray', 'size', 'first', 'initial', 'rest', 'last', 'without', 'indexOf',
+	'shuffle', 'lastIndexOf', 'isEmpty', 'groupBy'];
+
+// Mix in each Underscore method as a proxy to `Collection#models`.
+_.each(methods, function(method) {
+	Collection.prototype[method] = function() {
+	return _[method].apply(_, [this.models].concat(_.toArray(arguments)));
+	};
+});
+
+// Backbone.Router
+// -------------------
+
+// Routers map faux-URLs to actions, and fire events when routes are
+// matched. Creating a new one sets its `routes` hash, if not set statically.
+var Router = Backbone.Router = function(options) {
+	options || (options = {});
+	if (options.routes) this.routes = options.routes;
+	this._bindRoutes();
+	this.initialize.apply(this, arguments);
+};
+
+// Cached regular expressions for matching named param parts and splatted
+// parts of route strings.
+var namedParam    = /:\w+/g;
+var splatParam    = /\*\w+/g;
+var escapeRegExp  = /[-[\]{}()+?.,\\^$|#\s]/g;
+
+// Set up all inheritable **Backbone.Router** properties and methods.
+_.extend(Router.prototype, Events, {
+
+	// Initialize is an empty function by default. Override it with your own
+	// initialization logic.
+	initialize: function(){},
+
+	// Manually bind a single named route to a callback. For example:
+	//
+	//     this.route('search/:query/p:num', 'search', function(query, num) {
+	//       ...
+	//     });
+	//
+	route: function(route, name, callback) {
+	Backbone.history || (Backbone.history = new History);
+	if (!_.isRegExp(route)) route = this._routeToRegExp(route);
+	if (!callback) callback = this[name];
+	Backbone.history.route(route, _.bind(function(fragment) {
+		var args = this._extractParameters(route, fragment);
+		callback && callback.apply(this, args);
+		this.trigger.apply(this, ['route:' + name].concat(args));
+		Backbone.history.trigger('route', this, name, args);
+	}, this));
+	return this;
+	},
+
+	// Simple proxy to `Backbone.history` to save a fragment into the history.
+	navigate: function(fragment, options) {
+	Backbone.history.navigate(fragment, options);
+	},
+
+	// Bind all defined routes to `Backbone.history`. We have to reverse the
+	// order of the routes here to support behavior where the most general
+	// routes can be defined at the bottom of the route map.
+	_bindRoutes: function() {
+	if (!this.routes) return;
+	var routes = [];
+	for (var route in this.routes) {
+		routes.unshift([route, this.routes[route]]);
+	}
+	for (var i = 0, l = routes.length; i < l; i++) {
+		this.route(routes[i][0], routes[i][1], this[routes[i][1]]);
+	}
+	},
+
+	// Convert a route string into a regular expression, suitable for matching
+	// against the current location hash.
+	_routeToRegExp: function(route) {
+	route = route.replace(escapeRegExp, '\\$&')
+				.replace(namedParam, '([^\/]+)')
+				.replace(splatParam, '(.*?)');
+	return new RegExp('^' + route + '$');
+	},
+
+	// Given a route, and a URL fragment that it matches, return the array of
+	// extracted parameters.
+	_extractParameters: function(route, fragment) {
+	return route.exec(fragment).slice(1);
+	}
+
+});
+
+// Backbone.History
+// ----------------
+
+// Handles cross-browser history management, based on URL fragments. If the
+// browser does not support `onhashchange`, falls back to polling.
+var History = Backbone.History = function() {
+	this.handlers = [];
+	_.bindAll(this, 'checkUrl');
+};
+
+// Cached regex for cleaning leading hashes and slashes .
+var routeStripper = /^[#\/]/;
+
+// Cached regex for detecting MSIE.
+var isExplorer = /msie [\w.]+/;
+
+// Has the history handling already been started?
+History.started = false;
+
+// Set up all inheritable **Backbone.History** properties and methods.
+_.extend(History.prototype, Events, {
+
+	// The default interval to poll for hash changes, if necessary, is
+	// twenty times a second.
+	interval: 50,
+
+	// Gets the true hash value. Cannot use location.hash directly due to bug
+	// in Firefox where location.hash will always be decoded.
+	getHash: function(windowOverride) {
+	var loc = windowOverride ? windowOverride.location : window.location;
+	var match = loc.href.match(/#(.*)$/);
+	return match ? match[1] : '';
+	},
+
+	// Get the cross-browser normalized URL fragment, either from the URL,
+	// the hash, or the override.
+	getFragment: function(fragment, forcePushState) {
+	if (fragment == null) {
+		if (this._hasPushState || forcePushState) {
+		fragment = window.location.pathname;
+		var search = window.location.search;
+		if (search) fragment += search;
+		} else {
+		fragment = this.getHash();
+		}
+	}
+	if (!fragment.indexOf(this.options.root)) fragment = fragment.substr(this.options.root.length);
+	return fragment.replace(routeStripper, '');
+	},
+
+	// Start the hash change handling, returning `true` if the current URL matches
+	// an existing route, and `false` otherwise.
+	start: function(options) {
+	if (History.started) throw new Error("Backbone.history has already been started");
+	History.started = true;
+
+	// Figure out the initial configuration. Do we need an iframe?
+	// Is pushState desired ... is it available?
+	this.options          = _.extend({}, {root: '/'}, this.options, options);
+	this._wantsHashChange = this.options.hashChange !== false;
+	this._wantsPushState  = !!this.options.pushState;
+	this._hasPushState    = !!(this.options.pushState && window.history && window.history.pushState);
+	var fragment          = this.getFragment();
+	var docMode           = document.documentMode;
+	var oldIE             = (isExplorer.exec(navigator.userAgent.toLowerCase()) && (!docMode || docMode <= 7));
+
+	if (oldIE) {
+		this.iframe = $('<iframe src="javascript:0" tabindex="-1" />').hide().appendTo('body')[0].contentWindow;
+		this.navigate(fragment);
+	}
+
+	// Depending on whether we're using pushState or hashes, and whether
+	// 'onhashchange' is supported, determine how we check the URL state.
+	if (this._hasPushState) {
+		$(window).bind('popstate', this.checkUrl);
+	} else if (this._wantsHashChange && ('onhashchange' in window) && !oldIE) {
+		$(window).bind('hashchange', this.checkUrl);
+	} else if (this._wantsHashChange) {
+		this._checkUrlInterval = setInterval(this.checkUrl, this.interval);
+	}
+
+	// Determine if we need to change the base url, for a pushState link
+	// opened by a non-pushState browser.
+	this.fragment = fragment;
+	var loc = window.location;
+	var atRoot  = loc.pathname == this.options.root;
+
+	// If we've started off with a route from a `pushState`-enabled browser,
+	// but we're currently in a browser that doesn't support it...
+	if (this._wantsHashChange && this._wantsPushState && !this._hasPushState && !atRoot) {
+		this.fragment = this.getFragment(null, true);
+		window.location.replace(this.options.root + '#' + this.fragment);
+		// Return immediately as browser will do redirect to new url
+		return true;
+
+	// Or if we've started out with a hash-based route, but we're currently
+	// in a browser where it could be `pushState`-based instead...
+	} else if (this._wantsPushState && this._hasPushState && atRoot && loc.hash) {
+		this.fragment = this.getHash().replace(routeStripper, '');
+		window.history.replaceState({}, document.title, loc.protocol + '//' + loc.host + this.options.root + this.fragment);
+	}
+
+	if (!this.options.silent) {
+		return this.loadUrl();
+	}
+	},
+
+	// Disable Backbone.history, perhaps temporarily. Not useful in a real app,
+	// but possibly useful for unit testing Routers.
+	stop: function() {
+	$(window).unbind('popstate', this.checkUrl).unbind('hashchange', this.checkUrl);
+	clearInterval(this._checkUrlInterval);
+	History.started = false;
+	},
+
+	// Add a route to be tested when the fragment changes. Routes added later
+	// may override previous routes.
+	route: function(route, callback) {
+	this.handlers.unshift({route: route, callback: callback});
+	},
+
+	// Checks the current URL to see if it has changed, and if it has,
+	// calls `loadUrl`, normalizing across the hidden iframe.
+	checkUrl: function(e) {
+	var current = this.getFragment();
+	if (current == this.fragment && this.iframe) current = this.getFragment(this.getHash(this.iframe));
+	if (current == this.fragment) return false;
+	if (this.iframe) this.navigate(current);
+	this.loadUrl() || this.loadUrl(this.getHash());
+	},
+
+	// Attempt to load the current URL fragment. If a route succeeds with a
+	// match, returns `true`. If no defined routes matches the fragment,
+	// returns `false`.
+	loadUrl: function(fragmentOverride) {
+	var fragment = this.fragment = this.getFragment(fragmentOverride);
+	var matched = _.any(this.handlers, function(handler) {
+		if (handler.route.test(fragment)) {
+		handler.callback(fragment);
+		return true;
+		}
+	});
+	return matched;
+	},
+
+	// Save a fragment into the hash history, or replace the URL state if the
+	// 'replace' option is passed. You are responsible for properly URL-encoding
+	// the fragment in advance.
+	//
+	// The options object can contain `trigger: true` if you wish to have the
+	// route callback be fired (not usually desirable), or `replace: true`, if
+	// you wish to modify the current URL without adding an entry to the history.
+	navigate: function(fragment, options) {
+	if (!History.started) return false;
+	if (!options || options === true) options = {trigger: options};
+	var frag = (fragment || '').replace(routeStripper, '');
+	if (this.fragment == frag) return;
+
+	// If pushState is available, we use it to set the fragment as a real URL.
+	if (this._hasPushState) {
+		if (frag.indexOf(this.options.root) != 0) frag = this.options.root + frag;
+		this.fragment = frag;
+		window.history[options.replace ? 'replaceState' : 'pushState']({}, document.title, frag);
+
+	// If hash changes haven't been explicitly disabled, update the hash
+	// fragment to store history.
+	} else if (this._wantsHashChange) {
+		this.fragment = frag;
+		this._updateHash(window.location, frag, options.replace);
+		if (this.iframe && (frag != this.getFragment(this.getHash(this.iframe)))) {
+		// Opening and closing the iframe tricks IE7 and earlier to push a history entry on hash-tag change.
+		// When replace is true, we don't want this.
+		if(!options.replace) this.iframe.document.open().close();
+		this._updateHash(this.iframe.location, frag, options.replace);
+		}
+
+	// If you've told us that you explicitly don't want fallback hashchange-
+	// based history, then `navigate` becomes a page refresh.
+	} else {
+		window.location.assign(this.options.root + fragment);
+	}
+	if (options.trigger) this.loadUrl(fragment);
+	},
+
+	// Update the hash location, either replacing the current entry, or adding
+	// a new one to the browser history.
+	_updateHash: function(location, fragment, replace) {
+	if (replace) {
+		location.replace(location.toString().replace(/(javascript:|#).*$/, '') + '#' + fragment);
+	} else {
+		location.hash = fragment;
+	}
+	}
+});
+
+// Backbone.View
+// -------------
+
+// Creating a Backbone.View creates its initial element outside of the DOM,
+// if an existing element is not provided...
+var View = Backbone.View = function(options) {
+	this.cid = _.uniqueId('view');
+	this._configure(options || {});
+	this._ensureElement();
+	this.initialize.apply(this, arguments);
+	this.delegateEvents();
+};
+
+// Cached regex to split keys for `delegate`.
+var delegateEventSplitter = /^(\S+)\s*(.*)$/;
+
+// List of view options to be merged as properties.
+var viewOptions = ['model', 'collection', 'el', 'id', 'attributes', 'className', 'tagName'];
+
+// Set up all inheritable **Backbone.View** properties and methods.
+_.extend(View.prototype, Events, {
+
+	// The default `tagName` of a View's element is `"div"`.
+	tagName: 'div',
+
+	// jQuery delegate for element lookup, scoped to DOM elements within the
+	// current view. This should be prefered to global lookups where possible.
+	$: function(selector) {
+	return this.$el.find(selector);
+	},
+
+	// Initialize is an empty function by default. Override it with your own
+	// initialization logic.
+	initialize: function(){},
+
+	// **render** is the core function that your view should override, in order
+	// to populate its element (`this.el`), with the appropriate HTML. The
+	// convention is for **render** to always return `this`.
+	render: function() {
+	return this;
+	},
+
+	// Remove this view from the DOM. Note that the view isn't present in the
+	// DOM by default, so calling this method may be a no-op.
+	remove: function() {
+	this.$el.remove();
+	return this;
+	},
+
+	// For small amounts of DOM Elements, where a full-blown template isn't
+	// needed, use **make** to manufacture elements, one at a time.
+	//
+	//     var el = this.make('li', {'class': 'row'}, this.model.escape('title'));
+	//
+	make: function(tagName, attributes, content) {
+	var el = document.createElement(tagName);
+	if (attributes) $(el).attr(attributes);
+	if (content) $(el).html(content);
+	return el;
+	},
+
+	// Change the view's element (`this.el` property), including event
+	// re-delegation.
+	setElement: function(element, delegate) {
+	if (this.$el) this.undelegateEvents();
+	this.$el = (element instanceof $) ? element : $(element);
+	this.el = this.$el[0];
+	if (delegate !== false) this.delegateEvents();
+	return this;
+	},
+
+	// Set callbacks, where `this.events` is a hash of
+	//
+	// *{"event selector": "callback"}*
+	//
+	//     {
+	//       'mousedown .title':  'edit',
+	//       'click .button':     'save'
+	//       'click .open':       function(e) { ... }
+	//     }
+	//
+	// pairs. Callbacks will be bound to the view, with `this` set properly.
+	// Uses event delegation for efficiency.
+	// Omitting the selector binds the event to `this.el`.
+	// This only works for delegate-able events: not `focus`, `blur`, and
+	// not `change`, `submit`, and `reset` in Internet Explorer.
+	delegateEvents: function(events) {
+	if (!(events || (events = getValue(this, 'events')))) return;
+	this.undelegateEvents();
+	for (var key in events) {
+		var method = events[key];
+		if (!_.isFunction(method)) method = this[events[key]];
+		if (!method) throw new Error('Method "' + events[key] + '" does not exist');
+		var match = key.match(delegateEventSplitter);
+		var eventName = match[1], selector = match[2];
+		method = _.bind(method, this);
+		eventName += '.delegateEvents' + this.cid;
+		if (selector === '') {
+		this.$el.bind(eventName, method);
+		} else {
+		this.$el.delegate(selector, eventName, method);
+		}
+	}
+	},
+
+	// Clears all callbacks previously bound to the view with `delegateEvents`.
+	// You usually don't need to use this, but may wish to if you have multiple
+	// Backbone views attached to the same DOM element.
+	undelegateEvents: function() {
+	this.$el.unbind('.delegateEvents' + this.cid);
+	},
+
+	// Performs the initial configuration of a View with a set of options.
+	// Keys with special meaning *(model, collection, id, className)*, are
+	// attached directly to the view.
+	_configure: function(options) {
+	if (this.options) options = _.extend({}, this.options, options);
+	for (var i = 0, l = viewOptions.length; i < l; i++) {
+		var attr = viewOptions[i];
+		if (options[attr]) this[attr] = options[attr];
+	}
+	this.options = options;
+	},
+
+	// Ensure that the View has a DOM element to render into.
+	// If `this.el` is a string, pass it through `$()`, take the first
+	// matching element, and re-assign it to `el`. Otherwise, create
+	// an element from the `id`, `className` and `tagName` properties.
+	_ensureElement: function() {
+	if (!this.el) {
+		var attrs = getValue(this, 'attributes') || {};
+		if (this.id) attrs.id = this.id;
+		if (this.className) attrs['class'] = this.className;
+		this.setElement(this.make(this.tagName, attrs), false);
+	} else {
+		this.setElement(this.el, false);
+	}
+	}
+
+});
+
+// The self-propagating extend function that Backbone classes use.
+var extend = function (protoProps, classProps) {
+	var child = inherits(this, protoProps, classProps);
+	child.extend = this.extend;
+	return child;
+};
+
+// Set up inheritance for the model, collection, and view.
+Model.extend = Collection.extend = Router.extend = View.extend = extend;
+
+// Backbone.sync
+// -------------
+
+// Map from CRUD to HTTP for our default `Backbone.sync` implementation.
+var methodMap = {
+	'create': 'POST',
+	'update': 'PUT',
+	'delete': 'DELETE',
+	'read':   'GET'
+};
+
+// Override this function to change the manner in which Backbone persists
+// models to the server. You will be passed the type of request, and the
+// model in question. By default, makes a RESTful Ajax request
+// to the model's `url()`. Some possible customizations could be:
+//
+// * Use `setTimeout` to batch rapid-fire updates into a single request.
+// * Send up the models as XML instead of JSON.
+// * Persist models via WebSockets instead of Ajax.
+//
+// Turn on `Backbone.emulateHTTP` in order to send `PUT` and `DELETE` requests
+// as `POST`, with a `_method` parameter containing the true HTTP method,
+// as well as all requests with the body as `application/x-www-form-urlencoded`
+// instead of `application/json` with the model in a param named `model`.
+// Useful when interfacing with server-side languages like **PHP** that make
+// it difficult to read the body of `PUT` requests.
+Backbone.sync = function(method, model, options) {
+	var type = methodMap[method];
+
+	// Default options, unless specified.
+	options || (options = {});
+
+	// Default JSON-request options.
+	var params = {type: type, dataType: 'json'};
+
+	// Ensure that we have a URL.
+	if (!options.url) {
+	params.url = getValue(model, 'url') || urlError();
+	}
+
+	// Ensure that we have the appropriate request data.
+	if (!options.data && model && (method == 'create' || method == 'update')) {
+	params.contentType = 'application/json';
+	params.data = JSON.stringify(model.toJSON());
+	}
+
+	// For older servers, emulate JSON by encoding the request into an HTML-form.
+	if (Backbone.emulateJSON) {
+	params.contentType = 'application/x-www-form-urlencoded';
+	params.data = params.data ? {model: params.data} : {};
+	}
+
+	// For older servers, emulate HTTP by mimicking the HTTP method with `_method`
+	// And an `X-HTTP-Method-Override` header.
+	if (Backbone.emulateHTTP) {
+	if (type === 'PUT' || type === 'DELETE') {
+		if (Backbone.emulateJSON) params.data._method = type;
+		params.type = 'POST';
+		params.beforeSend = function(xhr) {
+		xhr.setRequestHeader('X-HTTP-Method-Override', type);
+		};
+	}
+	}
+
+	// Don't process data on a non-GET request.
+	if (params.type !== 'GET' && !Backbone.emulateJSON) {
+	params.processData = false;
+	}
+
+	// Make the request, allowing the user to override any Ajax options.
+	return $.ajax(_.extend(params, options));
+};
+
+// Wrap an optional error callback with a fallback error event.
+Backbone.wrapError = function(onError, originalModel, options) {
+	return function(model, resp) {
+	resp = model === originalModel ? resp : model;
+	if (onError) {
+		onError(originalModel, resp, options);
+	} else {
+		originalModel.trigger('error', originalModel, resp, options);
+	}
+	};
+};
+
+// Helpers
+// -------
+
+// Shared empty constructor function to aid in prototype-chain creation.
+var ctor = function(){};
+
+// Helper function to correctly set up the prototype chain, for subclasses.
+// Similar to `goog.inherits`, but uses a hash of prototype properties and
+// class properties to be extended.
+var inherits = function(parent, protoProps, staticProps) {
+	var child;
+
+	// The constructor function for the new subclass is either defined by you
+	// (the "constructor" property in your `extend` definition), or defaulted
+	// by us to simply call the parent's constructor.
+	if (protoProps && protoProps.hasOwnProperty('constructor')) {
+	child = protoProps.constructor;
+	} else {
+	child = function(){ parent.apply(this, arguments); };
+	}
+
+	// Inherit class (static) properties from parent.
+	_.extend(child, parent);
+
+	// Set the prototype chain to inherit from `parent`, without calling
+	// `parent`'s constructor function.
+	ctor.prototype = parent.prototype;
+	child.prototype = new ctor();
+
+	// Add prototype properties (instance properties) to the subclass,
+	// if supplied.
+	if (protoProps) _.extend(child.prototype, protoProps);
+
+	// Add static properties to the constructor function, if supplied.
+	if (staticProps) _.extend(child, staticProps);
+
+	// Correctly set child's `prototype.constructor`.
+	child.prototype.constructor = child;
+
+	// Set a convenience property in case the parent's prototype is needed later.
+	child.__super__ = parent.prototype;
+
+	return child;
+};
+
+// Helper function to get a value from a Backbone object as a property
+// or as a function.
+var getValue = function(object, prop) {
+	if (!(object && object[prop])) return null;
+	return _.isFunction(object[prop]) ? object[prop]() : object[prop];
+};
+
+// Throw an error when a URL is needed, and none is supplied.
+var urlError = function() {
+	throw new Error('A "url" property or function must be specified');
+};
+
+}).call(this);
