+++ test/apps/models/binding_listview/models/info.js
@@ -11,5 +11,21 @@ exports.definition = {
 			collection_name: 'info',
 			idAttribute: 'id'
 		}
+	},
+
+	extendCollection : function(Collection) {
+		_.extend(Collection.prototype, {
+
+			// For Backbone v1.1.2, uncomment this to override the fetch method
+			/*
+			fetch: function(options) {
+				options = options ? _.clone(options) : {};
+				options.reset = true;
+				return Backbone.Collection.prototype.fetch.call(this, options);
+			}
+			*/
+		});
+
+		return Collection;
 	}
 };
