+++ test/apps/models/todo/models/todo.js
@@ -39,6 +39,16 @@ exports.definition = {
 
     extendCollection : function(Collection) {
         _.extend(Collection.prototype, {
+
+			// For Backbone v1.1.2, uncomment this to override the fetch method
+			/*
+			fetch: function(options) {
+				options = options ? _.clone(options) : {};
+				options.reset = true;
+				return Backbone.Collection.prototype.fetch.call(this, options);
+			},
+			*/
+
             comparator: function(todo) {
 				return todo.get('done');
 			}
@@ -47,4 +57,3 @@ exports.definition = {
         return Collection;
     }
 };
-
